working solution - separate course component
